libgbm: refine gem handle flow
add gem handle map to trace the reference
The ioctl PRIME_FD_HANDLE may return the same handle by the same devicefd and dmabuf. If we destroy handle in each gbm_bo_destroy() will cause some handle in use to be closed.
To fix this, save the reference count of gem handle in the map and only close the handle when the reference count is 0.

Change-Id: Ib8b9ff14f9d6e26e84b7c34b1726a8c21d19c3b6